#1 Added database for saving Players #2 Refactoring of Engine class

+++ src/main/java/com/sfirsov/kalahgame/model/JoinGameRequest.java
@@ -1,14 +1,14 @@
 package com.sfirsov.kalahgame.model;
 
 public class JoinGameRequest {
-    int playerUid;
+    long playerUid;
     String opponentName;
 
-    public int getPlayerUid() {
+    public long getPlayerUid() {
         return playerUid;
     }
 
-    public void setPlayerUid(int playerUid) {
+    public void setPlayerUid(long playerUid) {
         this.playerUid = playerUid;
     }
